telegram_bot/requirements.txt
python-telegram-bot==21.5
aiohttp==3.10.5
Pillow==10.4.0
